Create curio.cards voting strategy 0.1.0
This is a strategy for Curio Cards, an early NFT trading cards
project from 2017.

Curio Cards predates the erc721 and erc1155 NFT specifications.
Each card is implemented as an erc20 with a max supply between
100-2000 tokens. This strategy weights the vote according to the
address's holdings of each card multiplied by the relative
scarcity of each card.

This might be one of the heavier strategies as it queries 30
different token contracts. The query is performed in parallel.